fixed expense deletion

+++ client/src/main/java/client/scenes/EventOverviewCtrl.java
@@ -15,7 +15,6 @@ import commons.Event;
 import commons.Expense;
 import commons.Participant;
 
-import jakarta.ws.rs.WebApplicationException;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
@@ -66,6 +65,8 @@ public class EventOverviewCtrl {
     private TableColumn<Expense, String> payerColumn;
     @FXML
     private TableColumn<Expense, LocalDate> dateColumn;
+    @FXML
+    private TableColumn<Expense, String> typeColumn;
 
     @FXML
     private Button editTitleButton;
@@ -144,6 +145,7 @@ public class EventOverviewCtrl {
         LanguageUtils.update(deleteExpenseButton, "deleteSelected");
 
 
+        LanguageUtils.update(typeColumn, "type1");
         LanguageUtils.update(titleColumn, "title");
         LanguageUtils.update(payerColumn, "payer");
         LanguageUtils.update(amountColumn, "amountEUR");
@@ -163,23 +165,26 @@ public class EventOverviewCtrl {
         amountColumn.setCellValueFactory(new PropertyValueFactory<>("amount"));
         payerColumn.setCellValueFactory(cellData -> new SimpleStringProperty(cellData.getValue().getPayer().getName()));
         dateColumn.setCellValueFactory(new PropertyValueFactory<>("localDate"));
+        typeColumn.setCellValueFactory(new PropertyValueFactory<>("type"));
 
         expensesTableView.setItems(observableExpenseList);
 
-        titleColumn.prefWidthProperty().bind(expensesTableView.widthProperty().divide(4).subtract(2));
-        amountColumn.prefWidthProperty().bind(expensesTableView.widthProperty().divide(4).subtract(1));
-        payerColumn.prefWidthProperty().bind(expensesTableView.widthProperty().divide(4).subtract(1));
-        dateColumn.prefWidthProperty().bind(expensesTableView.widthProperty().divide(4).subtract(1));
+        titleColumn.prefWidthProperty().bind(expensesTableView.widthProperty().divide(5).subtract(1));
+        amountColumn.prefWidthProperty().bind(expensesTableView.widthProperty().divide(5).subtract(1));
+        payerColumn.prefWidthProperty().bind(expensesTableView.widthProperty().divide(5).subtract(1));
+        dateColumn.prefWidthProperty().bind(expensesTableView.widthProperty().divide(5).subtract(1));
+        typeColumn.prefWidthProperty().bind(expensesTableView.widthProperty().divide(5).subtract(1));
 
         ScrollBar vScrollBar = (ScrollBar) expensesTableView.lookup(".scroll-bar:vertical");
 
         vScrollBar.visibleProperty().addListener((observable, wasVisible, isVisible) -> {
             if (isVisible) {
                 // If the scrollbar is visible, divide table width by 3 and subtract scrollbar width
-                titleColumn.prefWidthProperty().bind(expensesTableView.widthProperty().divide(4).subtract(6));
-                amountColumn.prefWidthProperty().bind(expensesTableView.widthProperty().divide(4).subtract(4));
-                payerColumn.prefWidthProperty().bind(expensesTableView.widthProperty().divide(4).subtract(4));
-                dateColumn.prefWidthProperty().bind(expensesTableView.widthProperty().divide(4).subtract(4));
+                titleColumn.prefWidthProperty().bind(expensesTableView.widthProperty().divide(5).subtract(4));
+                amountColumn.prefWidthProperty().bind(expensesTableView.widthProperty().divide(5).subtract(4));
+                payerColumn.prefWidthProperty().bind(expensesTableView.widthProperty().divide(5).subtract(4));
+                dateColumn.prefWidthProperty().bind(expensesTableView.widthProperty().divide(5).subtract(4));
+                typeColumn.prefWidthProperty().bind(expensesTableView.widthProperty().divide(5).subtract(4));
             }
         });
     }
@@ -238,7 +243,7 @@ public class EventOverviewCtrl {
 
                 MenuItem deleteMenuItem = new MenuItem(mainCtrl.getString("delete"));
                 deleteMenuItem.setOnAction(e -> {
-                    if (showConfirmationDialogExpenses(mainCtrl.getString("confirmationMessageDelete"))) {
+                    if (showConfirmationDialogExpenses(mainCtrl.getString("confirmationMessageDeleteExpense"))) {
                         deleteExpense();
                     }
                 });
@@ -277,7 +282,7 @@ public class EventOverviewCtrl {
 
                 MenuItem deleteMenuItem = new MenuItem(mainCtrl.getString("delete"));
                 deleteMenuItem.setOnAction(e -> {
-                    if (showConfirmationDialogParticipants(mainCtrl.getString("confirmationMessageDelete"))) {
+                    if (showConfirmationDialogParticipants(mainCtrl.getString("confirmationMessageDeleteParticipant"))) {
                         deleteParticipant();
                     }
                 });
@@ -328,9 +333,9 @@ public class EventOverviewCtrl {
     public void deleteParticipant() {
         if (selectedParticipant != null) {
             if(!selectedParticipant.getExpensesPaidBy().isEmpty() || !selectedParticipant.getExpensesToPay().isEmpty()) {
-                throw new WebApplicationException(mainCtrl.getString("invalidInput"));
+                showConfirmationDialogParticipants(mainCtrl.getString("errorDeleteParticipantWithExpenses"));
             }
-            else if(showConfirmationDialogParticipants(mainCtrl.getString("confirmationMessageDelete"))){
+            else if(showConfirmationDialogParticipants(mainCtrl.getString("confirmationMessageDeleteParticipant"))){
                 List<Debt> participantDebts = server.getDebtsByParticipant(selectedParticipant);
                 if(participantDebts != null && !participantDebts.isEmpty()) {
                     for (Debt d : participantDebts) {
@@ -371,7 +376,30 @@ public class EventOverviewCtrl {
     }
 
     public void deleteExpense() {
-        if (selectedExpense != null) {
+        if (selectedExpense != null && showConfirmationDialogParticipants(mainCtrl.getString("confirmationMessageDeleteExpense"))) {
+            double amountChange = selectedExpense.getAmount() * (-1.0);
+            for(Participant p : selectedExpense.getSplitters()) {
+                for(Debt debt : server.getDebtsByParticipant(p)){
+                    if(debt.getDebtor().getId() == p.getId() && debt.getCreditor().getId() == selectedExpense.getPayer().getId()) {
+                        debt.addAmount(amountChange / selectedExpense.getSplitters().size()); // dividing the amount equally between all the splitters
+                        if(debt.getAmount() < 0) {
+                            debt.setAmount(debt.getAmount() * (-1.0));
+                            debt.setDebtor(debt.getCreditor());
+                            debt.setCreditor(p);
+                        }
+                        server.editDebt(debt);
+                    }
+                    else if(debt.getCreditor().getId() == p.getId() && debt.getDebtor().getId() == selectedExpense.getPayer().getId()) {
+                        debt.addAmount((-1.0) * amountChange / selectedExpense.getSplitters().size());
+                        if(debt.getAmount() < 0) {
+                            debt.setAmount(debt.getAmount() * (-1.0));
+                            debt.setCreditor(debt.getDebtor());
+                            debt.setDebtor(p);
+                        }
+                        server.editDebt(debt);
+                    }
+                }
+            }
             server.deleteExpense(selectedExpense);
             initExpensesTableView(event);
         }
